Decreased overhead of using taint
Via using typing.TYPE_CHECKING, dynamically swap out the boxes for pseudo identity
functions (that are secretly also objects to satisfy the type checker). This means
when actually run, there is drastically less overhead. This also means that I
should be able to factor out the actual bodies of all the functions in the box
classes replacing them with ellipsis.